Move ColumnStat to File dir
